Make ResourceIdentifier non-final
A number of internal projects use ResourceIdentifier, which is a final class. This makes it much harder to use tools like Antipatterns, because you have to FBSuppressWarnings every single usage of ResourceIdentifier.

It'd be great if we could remove this restriction, as the advantages of it being final don't obviously outweigh this big negative.